refactor: クラスのインライン化
