--- dev/20190524/session/api/src/main/java/fizzbuzz/FizzBuzzData.java
@@ -1,17 +1,0 @@
-package fizzbuzz;
-
-import java.util.List;
-
-public class FizzBuzzData {
-  private List<String> _list;
-
-  public List<String> getList() {
-    return _list;
-  }
-
-  public void setList(List<String> list) {
-    _list = list;
-  }
-
-}
-
